adding gradle
