+++ .gitignore
@@ -1,3 +1,5 @@
-.idea/
-.gradle/
-build/
+# Ignore Gradle project-specific cache directory
+.gradle
+.idea
+# Ignore Gradle build output directory
+build
