+++ app/src/main/java/ru/clevertec/MyException.java
@@ -1,4 +1,4 @@
-package comp;
+package ru.clevertec;
 
 public class MyException extends Exception {
